backend/requirements.txt
Django==3.2.16
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
pytz==2022.7
sqlparse==0.4.3
python-dotenv==0.21.0
